chore: hoist test-helpers to be used by all packages
