+++ examples/with-react-todos/src/effects/toast.test.js
@@ -1,8 +1,8 @@
 import registerToastEffect, { toast } from './toast';
 import { store as storeModule } from 'reffects-store';
 import * as timerModule from '../infrastructure/timer';
-import { destroyAllMocks } from '../../test-helpers/fixtures';
-import { callsTo } from '../../test-helpers/mockHelpers';
+import { destroyAllMocks } from '../../../../test-helpers/fixtures';
+import { callsTo } from '../../../../test-helpers/mockHelpers';
 import { clearHandlers, getEffectHandler } from 'reffects';
 
 describe('toast effect', () => {
