--- examples/with-zustand-todos/test-helpers/fixtures.js
@@ -1,17 +1,0 @@
-import { clearHandlers } from 'reffects';
-
-export function applyEventsFixture(registerEventsCallback) {
-  beforeAll(() => {
-    registerEventsCallback();
-  });
-
-  afterAll(() => {
-    clearHandlers();
-    destroyAllMocks();
-  });
-}
-
-export function destroyAllMocks() {
-  jest.restoreAllMocks();
-  jest.clearAllMocks();
-}
