--- examples/with-zustand-todos/test-helpers/mockHelpers.js
@@ -1,3 +1,0 @@
-export function callsTo(spyFn) {
-  return spyFn.mock.calls;
-}
