--- packages/batteries/test-helpers/fixtures.js
@@ -1,17 +1,0 @@
-import { clearHandlers } from 'reffects';
-
-export function destroyAllMocks() {
-  jest.restoreAllMocks();
-  jest.clearAllMocks();
-}
-
-export function applyEventsFixture(eventsRegistrar) {
-  beforeAll(() => {
-    eventsRegistrar();
-  });
-
-  afterAll(() => {
-    clearHandlers();
-    destroyAllMocks();
-  });
-}
