+++ packages/reffects-store/src/batteries/state.test.js
@@ -1,7 +1,7 @@
 import { clearHandlers, coeffect, getCoeffectHandler } from 'reffects';
 import registerStateCoeffect, { stateGet } from './state';
-import { destroyAllMocks } from '../test-helpers/fixtures';
-import { callsTo } from '../test-helpers/mockHelpers';
+import { destroyAllMocks } from '../../../../test-helpers/fixtures';
+import { callsTo } from '../../../../test-helpers/mockHelpers';
 
 describe('state coeffect', () => {
   afterEach(() => {
