--- packages/reffects-store/src/test-helpers/mockHelpers.js
@@ -1,5 +1,0 @@
-export function callsTo(spyFn) {
-  return spyFn.mock.calls;
-}
-
-export default callsTo;
