--- packages/reffects/test-helpers/fixtures.js
@@ -1,4 +1,0 @@
-export function destroyAllMocks() {
-  jest.restoreAllMocks();
-  jest.clearAllMocks();
-}
