+++ test-helpers/mockHelpers.js
@@ -1,0 +1,7 @@
+function callsTo(spyFn) {
+  return spyFn.mock.calls;
+}
+
+module.exports = {
+  callsTo
+}
